[kcms/user] Set interactive auth flag for more calls
Creating and deleting users needs that too

This also means that we need to be more correct with the types passed into the DBus calls since with passing the args as list we don't get implicit conversion to the correct type any more

BUG: 450122
FIXED-IN: 5.24.2
(cherry picked from commit 79113d99e44ace405bd9211e2c0f73659ff94c0f)

+++ kcms/users/src/kcm.h
@@ -32,7 +32,7 @@ public:
     ~KCMUser() override;
 
     Q_SCRIPTABLE bool createUser(const QString &name, const QString &realName, const QString &password, bool admin);
-    Q_SCRIPTABLE bool deleteUser(int index, bool deleteHome);
+    Q_SCRIPTABLE bool deleteUser(qint64 index, bool deleteHome);
     // Grab the initials of a string
     Q_SCRIPTABLE QString initializeString(const QString &stringToGrabInitialsOf);
     Q_SCRIPTABLE QString plonkImageInTempfile(const QImage &image);
